fix: [PIE-2771]: mini policy set renderer to use hooks (#7697)
* fix: mini policy set renderer to use hooks

* fix: removed console log

* test: updated snaps

* fix: updated version
